updated after reviewing. lift togglbookmark function of bookmark button up
