+++ components/List/List.js
@@ -1,48 +1,16 @@
-import {
-  StyledImage,
-  StyledList,
-  StyledListItem,
-  StyledName,
-  StyledRating,
-  StyledPicture,
-  CardContainer,
-} from "./List.styles";
-import Link from "next/link";
-import BookmarkButton from "../BookmarkButton/BookmarkButton";
+import { StyledList } from "./List.styles";
+import ListCard from "./ListCard/ListCard";
 
-export default function List({ restaurants, onToggleBookmark, userInfos }) {
+export default function List({ restaurants, onToggleFavorite, userInfos }) {
   return (
     <StyledList>
       {restaurants.map((restaurant) => (
-        <CardContainer key={restaurant.id}>
-          <BookmarkButton
-            onToggleBookmark={onToggleBookmark}
-            id={restaurant.id}
-            restaurant={restaurant}
-            userInfos={userInfos}
-          />
-          <Link href={`/restaurants/${restaurant.id}`}>
-            <StyledListItem>
-              <StyledName>{restaurant.name}</StyledName>
-              <StyledPicture>
-                <StyledImage
-                  alt="a photo of the restaurant"
-                  src={restaurant.photos[0]}
-                  fill
-                  sizes="(min-width: 768px) 100vw"
-                  priority
-                />
-              </StyledPicture>
-              <p>{restaurant.cuisine}</p>
-              <StyledRating>
-                {restaurant.rating}{" "}
-                <span role="img" aria-label="star">
-                  ⭐️
-                </span>
-              </StyledRating>
-            </StyledListItem>
-          </Link>
-        </CardContainer>
+        <ListCard
+          onToggleFavorite={onToggleFavorite}
+          userInfos={userInfos}
+          restaurant={restaurant}
+          key={restaurant.id}
+        />
       ))}
     </StyledList>
   );
