+++ pages/restaurants/index.js
@@ -2,13 +2,13 @@ import List from "../../components/List/List";
 import { restaurants } from "../../lib/data";
 import Heading from "../../components/Heading/Heading";
 
-export default function Restaurants({ onToggleBookmark, userInfos }) {
+export default function Restaurants({ onToggleFavorite, userInfos }) {
   return (
     <>
       <Heading>Restaurants</Heading>
       <List
         restaurants={restaurants}
-        onToggleBookmark={onToggleBookmark}
+        onToggleFavorite={onToggleFavorite}
         userInfos={userInfos}
       />
     </>
